♻️ Refactor: Move price and product name formatting logic to format.js for improved reusability
