+++ src/main/resources/static/js/product/product.js
@@ -1,4 +1,4 @@
-import { formatPrice } from "/js/common/format.js";
+import { formatPriceElements } from "/js/common/format.js";
 
 document.addEventListener('DOMContentLoaded', function () {
   async function productDetail(productId) {
@@ -23,14 +23,7 @@ document.addEventListener('DOMContentLoaded', function () {
 
   // productDetail(productId);
 
-  /**
-   * 가격을 한국 원화(KRW) 형식으로 포맷팅
-   */
-  const priceElements = document.querySelectorAll(".price");
-  priceElements.forEach(function (priceElement) {
-    const price = parseInt(priceElement.getAttribute("data-price"));
-    priceElement.textContent = formatPrice(price);
-  })
+  formatPriceElements();
 
   /**
    * 장바구니 담기 상품 수량 증가/감소 버튼
